pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "paperverifier"
version = "0.1.0"
description = "Enterprise-grade research paper verification tool with multi-LLM provider support"
readme = "README.md"
license = "LicenseRef-Proprietary"
requires-python = ">=3.11"
authors = [
    { name = "PaperVerifier Team" },
]

dependencies = [
    # LLM SDKs
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    # Web & CLI
    "click>=8.1",
    "streamlit>=1.38.0",
    "rich>=13.0",
    # Document parsing
    "pdfplumber>=0.11.0",
    "pypdf>=4.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "pypandoc>=1.13",
    "markdown>=3.5",
    # GitHub & URL
    "gitpython>=3.1",
    "requests>=2.31",
    "aiohttp>=3.9",
    # Data & Config
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    # Security
    "keyring>=25.0",
    # Async
    "nest-asyncio>=1.6",
    # Logging & Observability
    "structlog>=24.1",
    # JSON parsing
    "json5>=0.9",
    # Utilities
    "tenacity>=8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "pre-commit>=3.7",
]

[project.scripts]
paperverifier = "paperverifier.cli:cli"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "PT"]
ignore = ["S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
